+++ web/documentserver-example/nodejs/app.js
@@ -1,4 +1,4 @@
-/**
+﻿/**
  *
  * (c) Copyright Ascensio System SIA 2023
  *
@@ -718,12 +718,12 @@ app.post('/track', async (req, res) => { // define a handler for tracking file c
         if (isSubmitForm) {
           // new file
           if (newFileName) {
-            correctName = req.DocManager.getCorrectName(`${fileUtility.getFileName(fileName, true)}
-            -form${downloadExt}`, userAddress);
+            correctName = req.DocManager.getCorrectName(
+              `${fileUtility.getFileName(fileName, true)}-form${downloadExt}`, userAddress);
           } else {
             const ext = fileUtility.getFileExtension(fileName);
-            correctName = req.DocManager.getCorrectName(`${fileUtility.getFileName(fileName, true)}
-            -form${ext}`, userAddress);
+            correctName = req.DocManager.getCorrectName(
+              `${fileUtility.getFileName(fileName, true)}-form${ext}`, userAddress);
           }
           forcesavePath = req.DocManager.storagePath(correctName, userAddress);
         } else {
@@ -948,7 +948,7 @@ app.get('/editor', (req, res) => { // define a handler for editing document
     if (!canEdit && mode === 'edit') {
       mode = 'view';
     }
-    const submitForm = mode === 'fillForms' && userid === 'uid-1' && !1;
+    const submitForm = mode === 'fillForms' && userid === 'uid-1';
 
     // file config data
     const argss = {